% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqrIS.R
\docType{package}
\name{cqrIS-package}
\alias{cqrIS-package}
\alias{_PACKAGE}
\title{cqrIS: Censored Quantile Regression with Induced Smoothing}
\description{
This package is an implementation of the estimation and inference procedure of a quantile
regression model which is based on a stochastic differential equation, with a smoothed
version based on induced smoothing technique. The parameters are obtained either by minimizing
an equivalent objective $l_1$ function of an estimating equation, or by solving an induced-smoothed
version of it. Two generalizations to handle functional data and recurrent event data respectively
are also implemented. For inference procedure, the multiplier bootstrap approach is used.
}
\references{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ZexiCAI/cqrIS}
  \item Report bugs at \url{https://github.com/ZexiCAI/cqrIS/issues}
}

}
\author{
\strong{Maintainer}: Zexi Cai \email{zxcai@link.cuhk.edu.hk}

Authors:
\itemize{
  \item Tony Sit \email{tonysit@sta.cuhk.edu.hk} [contributor]
}

}
