% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqrIS_main.R
\name{cqrIS}
\alias{cqrIS}
\title{Censored Quantile Regression with Induced Smoothing}
\usage{
cqrIS(Z, X, cen, init = TRUE, tau.max = 0.8, grid = 0.01,
  tol = 1e-04, diverge = 10000, maxiter = 200, boot = FALSE,
  nboot = 250)
}
\arguments{
\item{Z}{A numeric matrix. Covariates}

\item{X}{A numeric vector. Survival time, not log-transformed}

\item{cen}{A numeric vector. Censoring indicator, 1 for failure, 0 for censored}

\item{init}{Logical. Whether Peng and Huang (2008) estimates are used as initial guess, default: TRUE for yes}

\item{tau.max}{A numeric number. Upper bound of estimated quantile, default: 0.8}

\item{grid}{A numeric number. The grid size, default: 0.01}

\item{tol}{A numeric number. Norm of the estimate for the update to be considered convergent, default: 1e-4}

\item{diverge}{A numeric number. Norm of the estimate for the update to be considered divergent, default: 1e4}

\item{maxiter}{A numeric number. Maximum number of iteration, default: 200}

\item{boot}{Logical. Whether to perform variance estimation, default: FALSE}

\item{nboot}{A numeric number. The number of bootstrap samples if boot=TRUE, default: 250}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.ISest, a matrix giving the estimates for the parameters. If boot=FALSE, it will contain p columns
  giving the coefficient estimates where p is the dimension of the covariate; if boot=TRUE, it will contain
  2*p columns giving the coefficient estimates in the first p columns and the estimated standard errors in
  the remaining columns.
  \item idenLim, a numeric giving the identifiability limit for the data
}
}
\description{
This function fits a censored quantile regression model to the provided survival dataset.
The parameters are obtained by solving an induced-smoothed version of the estimating
equation used in Peng and Huang (2008).
}
\examples{
\donttest{
dat <- ordin.sam200.cen25.homo
res.cqrIS <- cqrIS(Z=dat[,-c(1:2)], X=dat[,1], cen=dat[,2])
}

}
\references{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
