% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqrIS_main.R
\name{cqrIS_}
\alias{cqrIS_}
\title{Censored Quantile Regression with Induced Smoothing (The Interface)}
\usage{
cqrIS_(Z, X, cen, xi = rep(1, length(X)), init, tau.max, grid, tol,
  diverge, maxiter)
}
\arguments{
\item{Z}{Covariates}

\item{X}{Survival time, not log-transformed}

\item{cen}{Censoring indicator, 1 for failure, 0 for censored}

\item{xi}{Weights for samples, default: a 1-vector}

\item{init}{Whether Peng and Huang (2008) estimates are used as initial guess}

\item{tau.max}{Upper bound of estimated quantile}

\item{grid}{The grid size}

\item{tol}{Norm of the estimate for the update to be considered convergent}

\item{diverge}{Norm of the estimate for the update to be considered divergent}

\item{maxiter}{Maximum number of iteration}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.ISest, the estimates for the parameters
  \item idenLim, the identifiability limit for the data
}
}
\description{
This function fits a censored quantile regression model to the provided survival dataset.
The parameters are obtained by solving an induced-smoothed version of the estimating
equation used in Peng and Huang (2008). It also provides an interface for variance estimation
via multiplier boostrap approach.
}
\references{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
