% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIS.R
\name{estjiang}
\alias{estjiang}
\title{A Functional Censored Quantile Regression Model}
\usage{
estjiang(Z, W, X, cen, tau.max = 0.8, grid = 0.01, diverge = 10000,
  boot = FALSE, nboot = 250)
}
\arguments{
\item{Z}{A numeric matrix. Non-functional covariates}

\item{W}{A numeric matrix. Transformed functional covariates}

\item{X}{A numeric vector. Survival time, not log-transformed}

\item{cen}{A numeric vector. Censoring indicator, 1 for failure, 0 for censored}

\item{tau.max}{A numeric number. Upper bound of estimated quantile, default: 0.8}

\item{grid}{A numeric number. The grid size, default: 0.01}

\item{diverge}{A numeric number. Norm of the estimate for the update to be considered divergent, default: 1e4}

\item{boot}{Logical. Whether to perform variance estimation, default: FALSE}

\item{nboot}{A numeric number. The number of bootstrap samples if boot=TRUE, default: 250}
}
\value{
This function returns a list of lists with each list containing three elements:
\itemize{
  \item beta.est, a matrix giving the estimates for the non-functional coefficients. If boot=FALSE, it will
  contain p columns giving the coefficient estimates where p is the dimension of the non-functional covariate;
  if boot=TRUE, it will contain 2*p columns giving the coefficient estimates in the first p columns and the
  estimated standard errors in the remaining columns.
  \item gamma.est, a matrix giving the estimates for the functional coefficients. If boot=FALSE, it will
  contain d columns giving the coefficient estimates where d is the dimension of the basis-transformed functional
  covariate; if boot=TRUE, it will contain 2*d columns giving the coefficient estimates in the first d columns
  and the estimated standard errors in the remaining columns.
  \item idenLim, a numeric giving the identifiability limit for the data
}
}
\description{
This function fits a censored quantile regression model to the provided dataset where
the covariates include both non-functional and functional types. The parameters
are obtained by minimizing an equivalent objective $l_1$ function of an estimating
equation.
}
\examples{
\donttest{
dat <- funct.sam200.v10.nb5.cen20
res.jiang <- estjiang(Z=dat[,c(3:4)], W=dat[,c(5:9)], X=dat[,1], cen=dat[,2])
}

}
\references{
Jiang, F., Cheng, Q., Yin, G. and Shen, H. (2020),
"Generalizing Quantile Regression for Counting Processes with Applications to Recurrent Events,"
\emph{Journal of the American Statistical Association}, \bold{115}, 931-944.
}
\seealso{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
