% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIS.R
\name{estjiang_}
\alias{estjiang_}
\title{A Functional Censored Quantile Regression Model (The Interface)}
\usage{
estjiang_(Z, W, X, cen, xi = rep(1, length(X)), tau.max, grid, diverge)
}
\arguments{
\item{Z}{Non-functional covariates}

\item{W}{Transformed functional covariates}

\item{X}{Survival time, not log-transformed}

\item{cen}{Censoring indicator, 1 for failure, 0 for censored}

\item{xi}{Weights for samples, default: a 1-vector}

\item{tau.max}{Upper bound of estimated quantile}

\item{grid}{The grid size}

\item{diverge}{Norm of the estimate for the update to be considered divergent}
}
\value{
This function returns a list of lists with each list containing three elements:
\itemize{
  \item beta.est, the estimates for the non-functional coefficients
  \item gamma.est, the estimates for the functional coefficients
  \item idenLim, the identifiability limit for the data
}
}
\description{
An interface to fit a censored quantile regression model to the provided dataset where
the covariates include both non-functional and functional types. The parameters
are obtained by minimizing an equivalent objective $l_1$ function of an estimating
equation. It also provides an interface for variance estimation via multiplier bootstrap.
}
\references{
Jiang, F., Cheng, Q., Yin, G. and Shen, H. (2020),
"Generalizing Quantile Regression for Counting Processes with Applications to Recurrent Events,"
\emph{Journal of the American Statistical Association}, \bold{115}, 931-944.
}
\seealso{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
