% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqrIS_main.R
\name{estPH}
\alias{estPH}
\title{Censored Quantile Regression}
\usage{
estPH(Z, X, cen, tau.max = 0.8, grid = 0.01, diverge = 10000,
  boot = FALSE, nboot = 250)
}
\arguments{
\item{Z}{A numeric matrix. Covariates}

\item{X}{A numeric vector. Survival time, not log-transformed}

\item{cen}{A numeric vector. Censoring indicator, 1 for failure, 0 for censored}

\item{tau.max}{A numeric number. Upper bound of estimated quantile, default: 0.8}

\item{grid}{A numeric number. The grid size, default: 0.01}

\item{diverge}{A numeric number. Norm of the estimate for the update to be considered divergent, default: 1e4}

\item{boot}{Logical. Whether to perform variance estimation, default: FALSE}

\item{nboot}{A numeric number. The number of bootstrap samples if boot=TRUE, default: 250}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.est, a matrix giving the estimates for the parameters. If boot=FALSE, it will contain p columns
  giving the coefficient estimates where p is the dimension of the covariate; if boot=TRUE, it will contain
  2*p columns giving the coefficient estimates in the first p columns and the estimated standard errors in
  the remaining columns.
  \item idenLim, a numeric giving the identifiability limit for the data
}
}
\description{
This function fits a censored quantile regression model to the provided survival dataset.
The parameters are obtained by minimizing an equivalent objective $l_1$ function of an
estimating equation.
}
\examples{
\donttest{
dat <- ordin.sam200.cen25.homo
res.PH <- estPH(Z=dat[,-c(1:2)], X=dat[,1], cen=dat[,2])
}

}
\references{
Peng, L. and Huang, Y. (2008),
"Survival Analysis with Quantile Regression Models,"
\emph{Journal of the American Statistical Association}, \bold{103}, 637-649.
}
\seealso{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
