% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqrIS_main.R
\name{estPH_}
\alias{estPH_}
\title{Censored Quantile Regression (The Interface)}
\usage{
estPH_(Z, X, cen, xi = rep(1, length(X)), tau.max, grid, diverge)
}
\arguments{
\item{Z}{Covariates}

\item{X}{Survival time, not log-transformed}

\item{cen}{Censoring indicator, 1 for failure, 0 for censored}

\item{xi}{Weights for samples, default: a 1-vector}

\item{tau.max}{Upper bound of estimated quantile}

\item{grid}{The grid size}

\item{diverge}{Norm of the estimate for the update to be considered divergent}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.est, the estimates for the parameters
  \item idenLim, the identifiability limit for the data
}
}
\description{
This function fits a censored quantile regression model to the provided survival dataset.
The parameters are obtained by minimizing an equivalent objective $l_1$ function of an
estimating equation. It also provides an interface for variance estimation via multiplier
bootstrap.
}
\references{
Peng, L. and Huang, Y. (2008),
"Survival Analysis with Quantile Regression Models,"
\emph{Journal of the American Statistical Association}, \bold{103}, 637-649.
}
\seealso{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
