% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurIS.R
\name{estsun}
\alias{estsun}
\title{Recurrent Event Modelling via Generalizing Quantile Regression for Counting Processes}
\usage{
estsun(T, x, ID, grids, U, L, R, link = "1", boot = FALSE,
  nboot = 250)
}
\arguments{
\item{T}{A numeric vector. Recurrent event time}

\item{x}{A numeric matrix. Covariates for samples}

\item{ID}{A numeric vector. ID of those who experience recurrent events}

\item{grids}{A numeric number. Number of grids}

\item{U}{A numeric number. Upper bound of estimation}

\item{L}{A numeric vector. Left censoring variable}

\item{R}{A numeric vector. Right censoring variable}

\item{link}{A character string. Link function, default: "1"}

\item{boot}{Logical. Whether to perform variance estimation, default: FALSE}

\item{nboot}{A numeric number. The number of bootstrap samples if boot=TRUE, default: 250}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.est, a matrix giving the estimates for the parameters. If boot=FALSE, it will contain p columns
  giving the coefficient estimates where p is the dimension of the covariate; if boot=TRUE, it will contain
  2*p columns giving the coefficient estimates in the first p columns and the estimated standard errors in
  the remaining columns.
  \item idenLim, a numeric giving the identifiability limit for the data
}
}
\description{
This function fits a recurrent event model to the provided dataset where the covariates
are related to the "expected frequency" via a generalized linear model. The parameters
are obtained by minimizing an equivalent objective $l_1$ function of an estimating
equation.
}
\examples{
\donttest{
evt <- recur[[1]]
sam <- recur[[2]]
res.estsun <- estsun(T=evt[,1], x=sam[,-c(1:2)], ID=evt[,2], grids=100, U=4, L=sam[,1], R=sam[,2])
}

}
\references{
Sun, X., Peng, L., Huang, Y. and Lai, H.J. (2016),
"Generalizing Quantile Regression for Counting Processes with Applications to Recurrent Events,"
\emph{Journal of the American Statistical Association}, \bold{111}, 145-156.
}
\seealso{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
