% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurIS.R
\name{estsun_}
\alias{estsun_}
\title{Recurrent Event Modelling via Generalizing Quantile Regression for Counting Processes (The Interface)}
\usage{
estsun_(T, x, ID, grids, U, L, R, ZETA = rep(1, dim(X)[1]),
  zeta = rep(1, dim(x)[1]), link)
}
\arguments{
\item{T}{Recurrent event time}

\item{x}{Covariates for samples}

\item{ID}{ID of those who experience recurrent events}

\item{grids}{Number of grids}

\item{U}{Upper bound of estimation}

\item{L}{Left censoring variable}

\item{R}{Right censoring variable}

\item{ZETA}{Weights for those who experience recurrent events, default: a 1-vector}

\item{zeta}{Weights for samples, default: a 1-vector}

\item{link}{Link function, default: "1"}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.est, the estimates for the parameters
  \item idenLim, the identifiability limit for the data
}
}
\description{
An interface to fit a recurrent event model to the provided dataset where the covariates
are related to the "expected frequency" via a generalized linear model. The parameters
are obtained by minimizing an equivalent objective $l_1$ function of an estimating
equation. It also provides an interface for variance estimation via multiplier bootstrap.
The codes are adapted from Prof. Limin Peng's website (with modifications).
}
\references{
Sun, X., Peng, L., Huang, Y. and Lai, H.J. (2016),
"Generalizing Quantile Regression for Counting Processes with Applications to Recurrent Events,"
\emph{Journal of the American Statistical Association}, \bold{111}, 145-156.
}
\seealso{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
