% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funct.sam200.v20.nb5.cen40}
\alias{funct.sam200.v20.nb5.cen40}
\title{Random Survival Data with Time-Varying Covariates}
\format{A list containing a matrix of 200 rows and 9 columns
\describe{
  \item{X}{Numerical values giving the observed survival time}
  \item{del}{Numerical values giving the censoring status. 1 = Failure, 0 = Censored}
  \item{Z1}{Numerical values giving the first covariate}
  \item{Z2}{Numerical values giving the second covariate}
  \item{bspl4.1}{Numerical values giving the first transformed covariate using cubic splines}
  \item{bspl4.2}{Numerical values giving the second transformed covariate using cubic splines}
  \item{bspl4.3}{Numerical values giving the third transformed covariate using cubic splines}
  \item{bspl4.4}{Numerical values giving the fourth transformed covariate using cubic splines}
  \item{bspl4.5}{Numerical values giving the fifth transformed covariate using cubic splines}
}}
\source{
Randomly generated data
}
\usage{
funct.sam200.v20.nb5.cen40
}
\description{
A dataset containing randomly generated survival data with time-varying covariates. Refer to the package vignette for details.
}
\keyword{datasets}
