% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIS.R
\name{funIS_}
\alias{funIS_}
\title{Adaptation of cqrIS Method to a Functional Censored Quantile Regression Model (The Interface)}
\usage{
funIS_(Z, W, X, cen, xi = rep(1, length(X)), tau.max, grid, tol, diverge,
  maxiter)
}
\arguments{
\item{Z}{Non-functional covariates}

\item{W}{Transformed functional covariates}

\item{X}{Survival time, not log-transformed}

\item{cen}{Censoring indicator, 1 for failure, 0 for censored}

\item{xi}{Weights for samples, default: a 1-vector}

\item{tau.max}{Upper bound of estimated quantile}

\item{grid}{The grid size}

\item{tol}{Norm of the estimate for the update to be considered convergent}

\item{diverge}{Norm of the estimate for the update to be considered divergent}

\item{maxiter}{Maximum number of iteration}
}
\value{
This function returns a list of lists with each list containing three elements:
\itemize{
  \item beta.ISest, the estimates for the non-functional coefficients
  \item gamma.ISest, the estimates for the functional coefficients
  \item idenLim, the identifiability limit for the data
}
}
\description{
An interface to fit a censored quantile regression model to the provided dataset where
the covariates include both non-functional and functional types. The parameters
are obtained by solving an induced-smoothed version of the estimating equation used in
Jiang et al. (2020). It also provides an interface for variance estimation via multiplier
boostrap approach.
}
\references{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
