% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ordin.sam100.cen40.hetero}
\alias{ordin.sam100.cen40.hetero}
\title{Random Ordinary Survival Data}
\format{A list containing a matrix of 100 rows and 5 columns
\describe{
  \item{X}{Numerical values giving the observed survival time}
  \item{delta}{Numerical values giving the censoring status. 1 = Failure, 0 = Censored}
  \item{Z0}{Numerical values representing the intercept}
  \item{Z1}{Numerical values giving the first covariate}
  \item{Z2}{Numerical values giving the second covariate}
}}
\source{
Randomly generated data
}
\usage{
ordin.sam100.cen40.hetero
}
\description{
A dataset containing randomly generated survival data. Refer to the package vignette for details.
}
\keyword{datasets}
