% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recur}
\alias{recur}
\title{Random Recurrent Event Data}
\format{A list containing a list of two matrices. One is of 592 rows and 2 columns, the other is of 100 rows and 6 columns.
\describe{
  \item{event}{A list containing a data matrix for the event times. Includes:
    \itemize{
      \item{T}{Numerical values giving the recurrent event time}
      \item{ID}{Numerical values giving the position of the sample that experience recurrent event}
    }
  }
  \item{sample}{A list containing a data matrix for the event times. Includes:
    \itemize{
      \item{L}{Numerical values giving the left censoring time}
      \item{R}{Numerical values giving the right censoring time}
      \item{x0}{Numerical values giving the intercept}
      \item{x1}{Numerical values giving the first covariate}
      \item{x2}{Numerical values giving the second covariate}
      \item{x3}{Numerical values giving the third covariate}
    }
  }
}}
\source{
Randomly generated data
}
\usage{
recur
}
\description{
A dataset containing randomly generated recurrent event data. Refer to the package vignette for details.
}
\keyword{datasets}
