% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurIS.R
\name{recurIS}
\alias{recurIS}
\title{Adaptation of cqrIS Method to a Recurrent Event Model}
\usage{
recurIS(T, x, ID, grids, U, L, R, link = "1", boot = FALSE,
  nboot = 250, tol = 1e-05, diverge = 10000, maxiter = 200)
}
\arguments{
\item{T}{A numeric vector. Recurrent event time}

\item{x}{A numeric matrix. Covariates for samples}

\item{ID}{A numeric vector. ID of those who experience recurrent events}

\item{grids}{A numeric number. Number of grids}

\item{U}{A numeric number. Upper bound of estimation}

\item{L}{A numeric vector. Left censoring variable}

\item{R}{A numeric vector. Right censoring variable}

\item{link}{A character string. Link function, default: "1"}

\item{boot}{Logical. Whether to perform variance estimation, default: FALSE}

\item{nboot}{A numeric number. The number of bootstrap samples if boot=TRUE, default: 250}

\item{tol}{A numeric number. Norm of the estimate for the update to be considered convergent, default: 1e-5}

\item{diverge}{A numeric number. Norm of the estimate for the update to be considered divergent, default: 1e4}

\item{maxiter}{A numeric number. Maximum number of iteration, default: 200}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.recurISest, a matrix giving the estimates for the parameters. If boot=FALSE, it will contain p columns
  giving the coefficient estimates where p is the dimension of the covariate; if boot=TRUE, it will contain
  2*p columns giving the coefficient estimates in the first p columns and the estimated standard errors in
  the remaining columns.
  \item idenLim, a numeric giving the identifiability limit for the data
}
}
\description{
This function fits a recurrent event model to the provided dataset where the covariates
are related to the "expected frequency" via a generalized linear model. The parameters
are obtained by solving an induced-smoothed version of the estimating equation used in
Sun et al. (2016).
}
\examples{
\donttest{
evt <- recur[[1]]
sam <- recur[[2]]
res.recurIS <- recurIS(T=evt[,1], x=sam[,-c(1:2)], ID=evt[,2], grids=100, U=4, L=sam[,1], R=sam[,2])
}

}
\references{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
