% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurIS.R
\name{recurIS_}
\alias{recurIS_}
\title{Adaptation of cqrIS Method to a Recurrent Event Model (The Interface)}
\usage{
recurIS_(T, x, ID, grids, U, L, R, zeta = rep(1, dim(x)[1]), link, tol,
  diverge, maxiter)
}
\arguments{
\item{T}{Recurrent event time}

\item{x}{Covariates for samples}

\item{ID}{ID of those who experience recurrent events}

\item{grids}{Number of grids}

\item{U}{Upper bound of estimation}

\item{L}{Left censoring variable}

\item{R}{Right censoring variable}

\item{zeta}{Weights for samples, default: a 1-vector}

\item{link}{Link function, default: "1"}

\item{tol}{Norm of the estimate for the update to be considered convergent}

\item{diverge}{Norm of the estimate for the update to be considered divergent}

\item{maxiter}{Maximum number of iteration}
}
\value{
This function returns a list of lists with each list containing two elements:
\itemize{
  \item beta.ISest, the estimates for the parameters
  \item idenLim, the identifiability limit for the data
}
}
\description{
An interface to fit a recurrent event model to the provided dataset where the covariates
are related to the "expected frequency" via a generalized linear model. The parameters
are obtained by solving an induced-smoothed version of the estimating equation used in
Sun et al. (2016). It also provides an interface for variance estimation via multiplier
boostrap approach.
}
\references{
Cai, Z. and Sit, T. (2020+),
"Censored Quantile Regression with Induced Smoothing,"
\emph{Working Paper}.
}
